refactor: split sendAnswer
